Condense all pervious updates to a single script
This will be skipped on machines which were installed this year (and
thus already shipped with version 92) or which subsequently installed
version 92 successfully.

This script is for older machines which failed to install a previous
update but which somehow got rebased to this repository. This will soon
be forced by the Updater4MC repository.

TODO: MailScanner patches could be simplified and made safer. Need to
test on:
pre-2025, pre-92 machine
pre-2025, post-92 machine
post-2025 (post-92 by default) machine

TODO: Build new images with this update (and rebase update for this
directory).

NOTE: Small change to `updater4mc.sh` causes all versions of this repo
to have '1' prependend to th eversion, so '00' will become '100'. This
will cause us to skip versions 93-100, but it is the easiest way to work
around the 2-digit versioning (V1C bug!).

NOTE: Keep (and re-number) update 92 (now 00) to still replace generic
SSH keys.